deplouyable and inventory implementation

+++ end-to-end-tests/script/CreateSmartCharacter.s.sol
@@ -17,14 +17,14 @@ contract CreateSmartCharacter is Script {
     StoreSwitch.setStoreAddress(worldAddress);
     // Load the private key from the `PRIVATE_KEY` environment variable (in .env)
     uint256 deployerPrivateKey = vm.envUint("PRIVATE_KEY");
-    
+
     // Test values for creating the smart character
     // TODO accept as parameters to the run method for test reproducability
     uint256 characterId = 1253;
-      // The address this character will be minted to
+    // The address this character will be minted to
     address characterAddress = address(0xf39Fd6e51aad88F6F4ce6aB8827279cffFb92266);
 
-    uint8 typeId = uint8(123);
+    uint256 typeId = 123;
     uint256 itemId = 234;
     uint256 volume = 100;
     string memory cid = "azerty";
@@ -32,13 +32,16 @@ contract CreateSmartCharacter is Script {
 
     // Start broadcasting transactions from the deployer account
     vm.startBroadcast(deployerPrivateKey);
-    SmartCharacterLib.World memory smartCharacter = SmartCharacterLib.World({iface: IBaseWorld(worldAddress), namespace: "frontier" });
-    
+    SmartCharacterLib.World memory smartCharacter = SmartCharacterLib.World({
+      iface: IBaseWorld(worldAddress),
+      namespace: "frontier"
+    });
+
     smartCharacter.createCharacter(
       characterId,
       characterAddress,
       EntityRecordTableData({ typeId: typeId, itemId: itemId, volume: volume }),
-      EntityRecordOffchainTableData({name: characterName, dappURL: "noURL", description: "."}),
+      EntityRecordOffchainTableData({ name: characterName, dappURL: "noURL", description: "." }),
       cid
     );
     vm.stopBroadcast();
