+++ mud-contracts/frontier-world/src/modules/entity-record/systems/EntityRecord.sol
@@ -23,7 +23,7 @@ contract EntityRecord is EveSystem {
   function createEntityRecord(
     uint256 entityId,
     uint256 itemId,
-    uint8 typeId,
+    uint256 typeId,
     uint256 volume
   ) public hookable(entityId, _systemId()) {
     EntityRecordTable.set(_namespace().entityRecordTableId(), entityId, itemId, typeId, volume);
@@ -71,10 +71,7 @@ contract EntityRecord is EveSystem {
    * @param entityId we create an off-chain record for
    * @param name name of that entity
    */
-  function setName(
-    uint256 entityId,
-    string memory name
-  ) public hookable(entityId, _systemId()) {
+  function setName(uint256 entityId, string memory name) public hookable(entityId, _systemId()) {
     EntityRecordOffchainTable.setName(_namespace().entityRecordOffchainTableId(), entityId, name);
   }
 
@@ -85,10 +82,7 @@ contract EntityRecord is EveSystem {
    * @param entityId we create an off-chain record for
    * @param dappURL link to that entity's dApp URL
    */
-  function setDappURL(
-    uint256 entityId,
-    string memory dappURL
-  ) public hookable(entityId, _systemId()) {
+  function setDappURL(uint256 entityId, string memory dappURL) public hookable(entityId, _systemId()) {
     EntityRecordOffchainTable.setDappURL(_namespace().entityRecordOffchainTableId(), entityId, dappURL);
   }
 
@@ -99,10 +93,7 @@ contract EntityRecord is EveSystem {
    * @param entityId we create an off-chain record for
    * @param description description of that entity
    */
-  function setDescription(
-    uint256 entityId,
-    string memory description
-  ) public hookable(entityId, _systemId()) {
+  function setDescription(uint256 entityId, string memory description) public hookable(entityId, _systemId()) {
     EntityRecordOffchainTable.setDescription(_namespace().entityRecordOffchainTableId(), entityId, description);
   }
